refactor(input): kebab-case
